+++ src/GameCuaTao/MarioSample/Models/Mario.h
@@ -9,7 +9,7 @@ namespace MarioSample
 	{
 	public:
 		void LoadContent(ContentManager &content) override;
-		void Update(unsigned long deltaTime) override;
+		void Update(float deltaTime) override;
 		void Draw(ISpriteBatch_ spriteBatch) override;
 
 	private:
